split stuffs to their own folders by role
