+++ handlers/handlers.go
@@ -1,0 +1,117 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"net/http"
+	"time"
+
+	"goweb/dbwrapper"
+	"goweb/models"
+	"goweb/transactions"
+
+	"github.com/gorilla/mux"
+)
+
+type DB dbwrapper.DB
+
+// type Tx transactions.Tx
+
+// Beginx starts and returns a new transaction
+func (db *DB) Beginx() (*transactions.Tx, error) {
+	tx, err := db.DB.Beginx()
+	if err != nil {
+		return nil, err
+	}
+	return &transactions.Tx{Tx: tx}, nil
+}
+
+func (db *DB) UserHandler() http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		vars := mux.Vars(r)
+		name := vars["name"]
+
+		tx, err := db.Beginx()
+		if err != nil {
+			log.Println(err)
+		}
+		tx.CreatePerson(&models.Person{FirstName: name})
+		tx.Commit()
+	})
+}
+
+func (db *DB) UserList() http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		var users []models.Person
+		err := db.Select(&users, "SELECT * FROM person")
+		if err != nil {
+			log.Println(err)
+			http.Error(w, err.Error(), 500)
+			return
+		}
+		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+		w.WriteHeader(http.StatusOK)
+		if err := json.NewEncoder(w).Encode(users); err != nil {
+			log.Println(err)
+			http.Error(w, err.Error(), 500)
+			return
+		}
+	})
+}
+
+func (db *DB) GenDataHandler() http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		tx, err := db.Beginx()
+		if err != nil {
+			log.Println(err)
+		}
+		tx.GenerateData()
+		if err := tx.Commit(); err != nil {
+			log.Println(err)
+		}
+	})
+}
+
+// my version copied from tsenart's, looks like more of a mess but it works!
+func WelcomeHandler(config string) func(w http.ResponseWriter, r *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		vars := mux.Vars(r)
+		name := vars["name"]
+		fmt.Fprintf(w, "Welcome!, config: %s, user: %s", config, name)
+	}
+}
+
+func Middleware(l *log.Logger, next func(w http.ResponseWriter, r *http.Request)) func(w http.ResponseWriter, r *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		began := time.Now()
+		next(w, r)
+		l.Printf("%s: %s %s took %s", time.Now(), r.Method, r.URL, time.Since(began))
+	}
+}
+
+// took from the awesome: https://gist.github.com/tsenart/5fc18c659814c078378d
+func UserHandler() http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		vars := mux.Vars(r)
+		name := vars["name"]
+		fmt.Fprintf(w, "user: %s", name)
+	})
+}
+
+func WithMetrics(l *log.Logger, next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		began := time.Now()
+		next.ServeHTTP(w, r)
+		l.Printf("%s: %s %s took %s", time.Now(), r.Method, r.URL, time.Since(began))
+	})
+}
+
+func About(w http.ResponseWriter, r *http.Request) {
+	m := models.Message{Text: "go API, build v0.0.001.992."}
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
+	if err := json.NewEncoder(w).Encode(m); err != nil {
+		log.Println(err)
+	}
+}
